requirements.txt
Appium-Python-Client==0.26
appnope==0.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
ddt==1.1.1
decorator==4.4.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Logbook==1.1.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyYAML==5.1.1
requests==2.22.0
selenium==3.0.2
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
xlrd==1.0.0
XlsxWriter==1.0.2
xlwt==1.3.0
